==> 03_BallTracking/2_TrackNet/3_Analyse_Validation_Results.ipynb <==
# Validation Results

In this notebook the results of the model training are loaded to compare against each other

import pandas as pd
import os

versions = [
    {"augmentation": True, "input_number": 3, "output_number": 3},
    {"augmentation": False, "input_number": 3, "output_number": 3},
    {"augmentation": True, "input_number": 3, "output_number": 1},
    {"augmentation": False, "input_number": 3, "output_number": 1},
    {"augmentation": True, "input_number": 5, "output_number": 3},
    {"augmentation": False, "input_number": 5, "output_number": 3},
    {"augmentation": True, "input_number": 5, "output_number": 5},
    {"augmentation": False, "input_number": 5, "output_number": 5}
]

all_results = []

for version in versions:
    exp_id = "TrackNet_"+ str(version["input_number"]) + "-in-" + str(version["output_number"]) + "-out_aug_" + str(version["augmentation"])
    results = pd.read_csv('exps_final/'+ exp_id + "/training_log.txt")
    
    # add version to results
    results['version'] = exp_id
    all_results.append(results)

all_results = pd.concat(all_results)

all_results.head()

# get the max f1-score for each version
max_f1 = all_results.groupby('version')['F1_Score'].max().reset_index()

# get the entire row for those values
all_results.merge(max_f1, on=['version', 'F1_Score'], how='inner')
