==> 03_BallTracking/2_TrackNet/5_Generate_Test_Metrics.ipynb <==
# Test Metrics

In this Notebook the metrcis for the differen TN models are calculated

### imports

import pandas as pd
import numpy as np
from scipy.spatial import distance

# load df
results = pd.read_csv('../results/Pre_Results_TN.csv')
results

### functions

# calculae metrics for row 
def calculate_metrics(row, threshold=4):
    predicted_position = row['predicted_position']
    actual_position = row['ground_truth_position']

    if actual_position == (-1,-1) and predicted_position == (-1,-1):
        # true negative
        return 0, 0, 1, 0
    elif actual_position == (-1,-1) and predicted_position != (-1,-1):
        # false positive
        return 0, 1, 0, 0
    elif actual_position != (-1,-1) and predicted_position == (-1,-1):
        # false negative
        return 0, 0, 0, 1
    else:
        if distance.euclidean(predicted_position, actual_position) < threshold:
            # true positive
            return 1, 0, 0, 0
        else:
            # false positive
            return 0, 1, 0, 0

# function to calculate complete metrics
def calculate_df_metrcis(frame_results_df, threshold=4):
    metrics_dict = {'Dataset_New': {'tp': 0, 'fp': 0, 'tn': 0, 'fn': 0},
                    'Dataset_TrackNet': {'tp': 0, 'fp': 0, 'tn': 0, 'fn': 0},
                    'Visibility_Visible': {'tp': 0, 'fp': 0, 'tn': 0, 'fn': 0},
                    'Visibility_Difficult': {'tp': 0, 'fp': 0, 'tn': 0, 'fn': 0},
                    'Visibility_Blurry': {'tp': 0, 'fp': 0, 'tn': 0, 'fn': 0},
                    'Visibility_Occluded': {'tp': 0, 'fp': 0, 'tn': 0, 'fn': 0},
                    'Visibility_None': {'tp': 0, 'fp': 0, 'tn': 0, 'fn': 0},
                    'Video_6': {'tp': 0, 'fp': 0, 'tn': 0, 'fn': 0},
                    'Total': {'tp': 0, 'fp': 0, 'tn': 0, 'fn': 0}}
    
    for i in range(len(frame_results_df)):
        row = frame_results_df.iloc[i]

        subset = "Dataset_" + row['subset']
        visibility = "Visibility_" + (row['points'][0]['visibility'] if  len(row['points']) > 0 else 'None')
        video = row['video']

        tp, fp, tn, fn = calculate_metrics(row, threshold)

        # subset metrics
        metrics_dict[subset]['tp'] += tp
        metrics_dict[subset]['fp'] += fp
        metrics_dict[subset]['tn'] += tn
        metrics_dict[subset]['fn'] += fn

        # visibility metrics
        metrics_dict[visibility]['tp'] += tp
        metrics_dict[visibility]['fp'] += fp
        metrics_dict[visibility]['tn'] += tn
        metrics_dict[visibility]['fn'] += fn

        #check if video is Video_1
        if video == 'Video_6':
            metrics_dict['Video_6']['tp'] += tp
            metrics_dict['Video_6']['fp'] += fp
            metrics_dict['Video_6']['tn'] += tn
            metrics_dict['Video_6']['fn'] += fn

        # total metrics
        metrics_dict['Total']['tp'] += tp
        metrics_dict['Total']['fp'] += fp
        metrics_dict['Total']['tn'] += tn
        metrics_dict['Total']['fn'] += fn
    
    return metrics_dict

def calculate_agg_metrics(tp, fp, tn, fn):
    # Calculate metrics
    accuracy = (tp + tn) / (tp + tn + fp + fn) if (tp + tn + fp + fn) > 0 else 0
    precision = tp / (tp + fp) if (tp + fp) > 0 else 0
    recall = tp / (tp + fn) if (tp + fn) > 0 else 0
    f1 = (2 * precision * recall) / (precision + recall) if (precision + recall) > 0 else 0

    metrics = {'Accuracy': accuracy, 'F1 Score': f1, 'Precision': precision, 'Recall': recall}
    return metrics

def calculate_positioning_error(frame_results_df, threshold=4):
    positioning_error = []
    for i in range(len(frame_results_df)):
        row = frame_results_df.iloc[i]
        predicted_position = row['predicted_position']
        actual_position = row['ground_truth_position']

        if actual_position != (-1,-1) and predicted_position != (-1,-1):
            error = distance.euclidean(predicted_position, actual_position)
            positioning_error.append(error)

    # get mean PE
    mean_positioning_error = np.mean(positioning_error)

    # get mean PE below TH
    mean_positioning_error_below_th = np.mean([pe for pe in positioning_error if pe < threshold])

    # count bins for 0, 1,2,3,4,1000
    bins = [0, 1, 2, 3, 4, 5, 10, 100000]

    # create counts
    counts = np.bincount(np.digitize(positioning_error, bins))
    
    return counts, mean_positioning_error, mean_positioning_error_below_th

## Calculate Metrics

# loop over rows and add f1 metrics
results['Positioning_Error_Counts'] = None  # Initialize the column
results['Positioning_Error_Counts'] = results['Positioning_Error_Counts'].astype(object)
for i in range(len(results)):
    frame_results = results.iloc[i]['frame_results']

    frame_results_df = pd.DataFrame(eval(frame_results))

    # calculate metrics
    metrics_dict = calculate_df_metrcis(frame_results_df)

    # calculate aggregated metrics
    agg_metrics = {}
    for key, value in metrics_dict.items():
        tp = value['tp']
        fp = value['fp']
        tn = value['tn']
        fn = value['fn']
        
        agg_metrics[key] = calculate_agg_metrics(tp, fp, tn, fn)

    # filter for only f1-score
    agg_metrics_f1 = {key: value['F1 Score'] for key, value in agg_metrics.items()}

    # convert to single columns
    for key, value in agg_metrics_f1.items():
        results.at[i, key] = value

    # calculate positioning error
    counts, mean_positioning_error, mean_positioning_error_below_th = calculate_positioning_error(frame_results_df)

    results.at[i, 'Positioning_Error_mean'] = mean_positioning_error
    results.at[i, 'Positioning_Error_mean_Below_Threshold'] = mean_positioning_error_below_th

    #add counts as a an object to the row
    results.at[i, 'Positioning_Error_Counts'] = counts.tolist()


# remove column frame_results
results = results.drop(columns=['frame_results'])

# Convert the 'version' column from string representation of dictionaries to actual dictionaries
results['version'] = results['version'].apply(eval)  # Use eval to convert strings to dictionaries

# Expand the dictionary into separate columns
version_df = pd.json_normalize(results['version'])

# Drop the original 'version' column and concatenate the expanded columns
results = pd.concat([results.drop(columns=['version']), version_df], axis=1)


# create a new version column that has a concat string
results['Model'] = "TrackNet " + results['input_number'].astype(str) + "-in-"+results['output_number'].astype(str) + "-out aug-"+results['augmentation'].astype(str)
#put version column in front
results = results[['Model'] + [col for col in results.columns if col != 'Model']]

# rename and filter columns
rename_dict = {
    'Model': 'Model',
    'f1': 'F1-Score Total',
    'precision': 'Precision Total',
    'recall': 'Recall Total',
    'Dataset_New': 'F1 Dataset New',
    'Dataset_TrackNet': 'F1 Dataset TrackNet',
    'Visibility_Visible': 'F1 Visibility Visibile',
    'Visibility_Difficult': 'F1 Visibility Difficult',
    'Visibility_Occluded': 'F1 Visibility Occluded',
    'Visibility_Blurry': 'F1 Visibility Blurry',
    'fps': 'FPS',
    'Positioning_Error_mean': 'PE Mean',
    'Positioning_Error_mean_Below_Threshold': 'PE Mean Below Threshold',
    'Positioning_Error_Counts': 'PE Counts',
}

results = results.rename(columns=rename_dict)

# filter columns
results = results[['Model', 'F1-Score Total', 'Precision Total', 'Recall Total', 'F1 Dataset New', 'F1 Dataset TrackNet', 'F1 Visibility Visibile', 'F1 Visibility Difficult', 'F1 Visibility Occluded', 'F1 Visibility Blurry', 'FPS', 'PE Mean', 'PE Mean Below Threshold', 'PE Counts']]
results

results.to_csv('../results/Results_TN.csv', index=False)
